Update output, switch to double

+++ src/datasets.cpp
@@ -3,7 +3,7 @@
 #include <fstream>
 #include <sstream>
 
-Dataset1D::Dataset1D(std::vector<std::vector<float>> data, std::vector<int> labels) {
+Dataset1D::Dataset1D(std::vector<std::vector<double>> data, std::vector<int> labels) {
     this->train_data = data;
     this->train_labels = labels;
 
@@ -44,7 +44,7 @@ void Dataset1D::load_data() {
 
     // Load training data
     for (int i = 0; i < this->train_size; i++) {
-        std::vector<float> row;
+        std::vector<double> row;
         std::string cell;
         std::getline(train_data_file, line);
         std::stringstream line_stream(line);
@@ -68,7 +68,7 @@ void Dataset1D::load_data() {
 
     // Load test data
     for (int i = 0; i < this->test_size; i++) {
-        std::vector<float> row;
+        std::vector<double> row;
         std::string cell;
         std::getline(test_data_file, line);
         std::stringstream line_stream(line);
